+++ cs25server/src/main/kotlin/edu/dongguk/cs25server/service/OrderApplicationService.kt
@@ -19,54 +19,5 @@ class OrderApplicationService(
     private val orderApplicationRepository: OrderApplicationRepository,
     private val itemHQRepository: ItemHQRepository
 ) {
-    // 입고 관리 발주 목록 조회
-    fun readOrderRequest(
-        itemName: String,
-        category: String?,
-        supplier: String?
-    ): ListResponseDto<List<OrderResponseDto>> {
-        val itemHQs: List<ItemHQ> = itemHQRepository.findAllByItemNameAndCategoryAndSupplier(
-            itemName,
-            ItemCategory.getCategory(category), Supplier.getSupplier(supplier)
-        )
 
-        val orderResponseDtos: List<OrderResponseDto> =
-            itemHQs.map(ItemHQ::toOrderResponse).toList()
-        return ListResponseDto(
-            datalist = orderResponseDtos,
-            page_info = PageInfo(
-                page = 0,
-                size = 0,
-                total_elements = orderResponseDtos.size,
-                total_pages = ceil(orderResponseDtos.size / 10.0).toInt()
-            )
-        )
-    }
-
-    // 재고 조회
-    fun readOrderStocks(
-        lack: Int,
-        itemName: String,
-        category: String?,
-        supplier: String?
-    ): ListResponseDto<List<OrderStockResponseDto>> {
-        val itemHQs: List<ItemHQ> = if (lack == 1) {
-            itemHQRepository.findAllByStockIsLessThanOrderSum()
-        } else {
-            itemHQRepository.findAllByItemNameAndCategoryAndSupplier(
-                itemName, ItemCategory.getCategory(category), Supplier.getSupplier(supplier)
-            )
-        }
-
-        val orderStockResponseDtos: List<OrderStockResponseDto> = itemHQs.map(ItemHQ::toOrderStockResponse).toList()
-        return ListResponseDto(
-            datalist = orderStockResponseDtos,
-            page_info = PageInfo(
-                page = 0,
-                size = 0,
-                total_elements = orderStockResponseDtos.size,
-                total_pages = ceil(orderStockResponseDtos.size / 10.0).toInt()
-            )
-        )
-    }
 }
